automatizacion para eliminar condolencias y cursos expirados

+++ admin/editar_solicitud.php
@@ -54,14 +54,14 @@ if ($_SERVER["REQUEST_METHOD"] == "POST") {
         $stmt->execute([$tipo, $idSolicitud]);
 
         // Redirigir después de editar
-        echo "<script>window.location.href='../public/solicitudes.php';</script>";
+        echo "<script>window.location.href='../admin/solicitudes.php';</script>";
         exit();
     } catch (Exception $e) {
         echo "Error: " . $e->getMessage();
     }
 } else {
     // Si no se recibieron datos por POST, redirigir a la página de lista de solicitudes
-    echo "<script>window.location.href='../public/solicitudes.php';</script>";
+    echo "<script>window.location.href='../admin/solicitudes.php';</script>";
     exit();
 }
 
@@ -101,7 +101,7 @@ function eliminarArchivoAntiguo($idSolicitud, $nombreUsuario, $directorioDestino
 
     $rutaArchivoEliminar = $directorioDestino . $nombreUsuario . "/" . $nombreArchivoEliminar;
 
-    if (file_exists($rutaArchivoEliminar)) {
+    if (file_exists($rutaArchivoEliminar) && is_file($rutaArchivoEliminar)) {
         unlink($rutaArchivoEliminar);
     }
 }
@@ -118,7 +118,7 @@ function eliminarArchivoYActualizarBD($idSolicitud, $valor, $tipo, $descripcion)
 
     $rutaArchivoEliminar = "../uploads/documentos/solicitudes/" . $nombreUsuario . "/" . $nombreArchivoEliminar;
 
-    if (file_exists($rutaArchivoEliminar)) {
+    if (file_exists($rutaArchivoEliminar) && is_file($rutaArchivoEliminar)) {
         unlink($rutaArchivoEliminar);
     }
