+++ publicista/editarDeportista.php
@@ -97,7 +97,7 @@ function eliminarArchivoAntiguo($idDeportista, $directorioDestino) {
 
     $rutaArchivoEliminar = $directorioDestino . $nombreArchivoEliminar;
 
-    if (file_exists($rutaArchivoEliminar)) {
+    if (file_exists($rutaArchivoEliminar) && is_file($rutaArchivoEliminar)) {
         unlink($rutaArchivoEliminar);
     }
 }
@@ -112,7 +112,7 @@ function eliminarArchivoYActualizarBD($idDeportista, $nombre, $deporte) {
 
     $rutaArchivoEliminar = "../uploads/deportes/deportistas/" . $nombreArchivoEliminar;
 
-    if (file_exists($rutaArchivoEliminar)) {
+    if (file_exists($rutaArchivoEliminar) && is_file($rutaArchivoEliminar)) {
         unlink($rutaArchivoEliminar);
     }
